# specs/001-ecom-crud-roadmap/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Ecom Backend API
  version: 1.0.0
  description: |
    Contract for 6-phase e-commerce backend roadmap.
    Phase 1 (CRUD) is fully implemented.
servers:
  - url: /api/v1
...
paths:
...
  /users:
    get:
      tags: [Users]
      summary: List all users
      operationId: listUsers
      responses:
        '200':
          description: User list returned
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponse'
    post:
      tags: [Users]
      summary: Create user
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreateRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Validation error or duplicate email
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      tags: [Users]
      summary: Get user by id
      operationId: getUser
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Users]
      summary: Update user (full replacement)
      operationId: updateUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateRequest'
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Users]
      summary: Delete user
      operationId: deleteUser
      responses:
        '204':
          description: User deleted
        '404':
          $ref: '#/components/responses/NotFound'

  # ── Products ────────────────────────────────────────────────────────────────
...
components:
  parameters:
    Id:
      in: path
      name: id
      required: true
      schema:
        type: string
        format: uuid
      description: Resource UUID
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    BadRequest:
      description: Validation error or business rule violation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
          example: NOT_FOUND
        message:
          type: string
          example: User with id '...' not found
        errors:
          type: array
          items:
            $ref: '#/components/schemas/FieldError'
        timestamp:
          type: string
          format: date-time
        correlation_id:
          type: string
          format: uuid
    FieldError:
      type: object
      properties:
        field:
          type: string
        message:
          type: string

    # ── User ────────────────────────────────────────────────────────────────
    UserCreateRequest:
      type: object
      required: [name, email, password, role]
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 120
          example: Alice Smith
        email:
          type: string
          format: email
          example: alice@example.com
        password:
          type: string
          minLength: 8
          example: s3cr3t!Pass
        role:
          type: string
          enum: [ADMIN, OPERATOR]
          example: OPERATOR
    UserUpdateRequest:
      type: object
      required: [name, email, role]
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 120
        email:
          type: string
          format: email
        role:
          type: string
          enum: [ADMIN, OPERATOR]
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
          enum: [ADMIN, OPERATOR]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    # ── Product ─────────────────────────────────────────────────────────────
...